// SIC/SIC Admin/src/types/seasonalTopics.ts
export type SeasonalLessonsDataArr = {
	id: number;
	uuid: string;
	name: string;
	namePinging: string;
	nameTraditionalChinese: string;
	nameSimplifiedChinese: string;
	vocabulary: string;
	vocabularyPinyin: string;
	vocabularyTraditionalChinese: string;
	vocabularySimplifiedChinese: string;
	sentences: string;
	sentencesPinyin: string;
	sentencesTraditionalChinese: string;
	sentencesSimplifiedChinese: string;
	order: number;
	rewardStars: number;
	isActive: boolean;
	topicId: string;
};

export type AddEditSeasonalLesson = {
	onClose: () => void;
	onSubmit: () => void;
	editData: SeasonalLessonsDataArr | null;
};